testing for spatie went well

+++ app/Http/Controllers/CategoryController.php
@@ -3,10 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['role:ADM|WHS|USR']);
+    }
     /**list all categories */
     public function index()
     {
@@ -19,14 +24,19 @@ class CategoryController extends Controller
         $fields = $request->validate([
             'name' => 'required|string|unique:categories,name',
         ]);
-        $category = Category::create([
-            'name' => $fields['name']
-        ]);
 
-        return [
-            'message' => 'category added',
-            'category' => $category
-        ];
+        if ($request->user()->can('category:register')) {
+            $category = Category::create([
+                'name' => $fields['name']
+            ]);
+
+            return [
+                'message' => 'category added',
+                'category' => $category
+            ];
+        } else {
+            return ['message' => 'unauthorised for this action'];
+        }
     }
 
     /** update category */
@@ -35,14 +45,19 @@ class CategoryController extends Controller
         $fields = $request->validate([
             'name' => 'required|string',
         ]);
-        $category = Category::find($id);
-        $category->name = $fields['name'];
-        $category->update();
 
-        return [
-            'message' => 'category updated',
-            'category' => $category
-        ];
+        if ($request->user()->can('category:update')) {
+            $category = Category::find($id);
+            $category->name = $fields['name'];
+            $category->update();
+
+            return [
+                'message' => 'category updated',
+                'category' => $category
+            ];
+        } else {
+            return ['message' => 'unathorised for this action'];
+        }
     }
 
     /** search category by id */
@@ -54,16 +69,23 @@ class CategoryController extends Controller
     /** delete category */
     public function destroy($id)
     {
-        $category = Category::find($id);
-        $response = $category->delete();
-        if ($response == 1) {
-            return [
-                'message' => 'category deleted',
-            ];
+        $user = User::find(auth()->user()->id);
+
+        if ($user->can('category:delete')) {
+            $category = Category::find($id);
+            $response = $category->delete();
+
+            if ($response == 1) {
+                return [
+                    'message' => 'category deleted',
+                ];
+            } else {
+                return [
+                    'message' => 'not deleted',
+                ];
+            }
         } else {
-            return [
-                'message' => 'not deleted',
-            ];
+            return ['message' => 'unaithorised for this action'];
         }
     }
 }
