+++ app/Http/Controllers/OrderItemController.php
@@ -4,10 +4,16 @@ namespace App\Http\Controllers;
 
 use App\Models\Order;
 use App\Models\OrderItem;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class OrderItemController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['role:ADM|WHS']);
+    }
+
     /**list all orderItems */
     public function index()
     {
@@ -24,35 +30,39 @@ class OrderItemController extends Controller
             'total_price' => 'nullable',
         ]);
 
-        $orderItem = new OrderItem();
-        $orderItem->product_id = $fields['product_id'];
-        $orderItem->order_id =  $fields['order_id'];
-        $orderItem->quantity =  $fields['quantity'];
+        if ($request->user()->can('orderItem:register')) {
+            $orderItem = new OrderItem();
+            $orderItem->product_id = $fields['product_id'];
+            $orderItem->order_id =  $fields['order_id'];
+            $orderItem->quantity =  $fields['quantity'];
 
-        $unit_price = $orderItem->product->unit_price;
+            $unit_price = $orderItem->product->unit_price;
 
-        $total_price = floatval($fields['quantity']) * $unit_price;
-        $orderItem->total_price =  $total_price;
+            $total_price = floatval($fields['quantity']) * $unit_price;
+            $orderItem->total_price =  $total_price;
 
-        $orderItem->save();
+            $orderItem->save();
 
-        $allItems = OrderItem::where('order_id', $orderItem->order_id)->get();
+            $allItems = OrderItem::where('order_id', $orderItem->order_id)->get();
 
-        $sum = 0;
-        foreach ($allItems as $item) {
-            $sum = $sum + $item->total_price;
+            $sum = 0;
+            foreach ($allItems as $item) {
+                $sum = $sum + $item->total_price;
+            }
+
+            $order = Order::find($orderItem->order_id);
+            $order->total = $sum;
+
+            $order->update();
+
+            return [
+                'message' => 'order item created',
+                'order item' => $orderItem,
+                'grand total' => $order
+            ];
+        } else {
+            return ['message' => 'unauthorised for this action'];
         }
-
-        $order = Order::find($orderItem->order_id);
-        $order->total = $sum;
-
-        $order->update();
-        
-        return [
-            'message' => 'order item created',
-            'order item' => $orderItem,
-            'grand total' => $order
-        ];
     }
 
     /** update order */
@@ -64,18 +74,21 @@ class OrderItemController extends Controller
             'quantity' => 'required',
             'total_price' => 'required',
         ]);
+        if ($request->user()->can('orderItem:update')) {
+            $orderItem = OrderItem::find($id);
+            $orderItem->product_id = $fields['product_id'];
+            $orderItem->order_id = $fields['order_id'];
+            $orderItem->quantity = $fields['quantity'];
+            $orderItem->total_price = $fields['total_price'];
+            $orderItem->update();
 
-        $orderItem = OrderItem::find($id);
-        $orderItem->product_id = $fields['product_id'];
-        $orderItem->order_id = $fields['order_id'];
-        $orderItem->quantity = $fields['quantity'];
-        $orderItem->total_price = $fields['total_price'];
-        $orderItem->update();
-
-        return [
-            'message' => 'order updated',
-            'product' => $orderItem
-        ];
+            return [
+                'message' => 'order updated',
+                'product' => $orderItem
+            ];
+        } else {
+            return ['message' => 'unauthorised for this action'];
+        }
     }
 
     /** search order by id */
@@ -87,16 +100,22 @@ class OrderItemController extends Controller
     /** delete order */
     public function destroy($id)
     {
-        $orderItem = OrderItem::find($id);
-        $response = $orderItem->delete();
-        if ($response == 1) {
-            return [
-                'message' => 'order item deleted',
-            ];
+        $user = User::find(auth()->user()->id);
+
+        if ($user->can('orderItem:delete')) {
+            $orderItem = OrderItem::find($id);
+            $response = $orderItem->delete();
+            if ($response == 1) {
+                return [
+                    'message' => 'order item deleted',
+                ];
+            } else {
+                return [
+                    'message' => 'not deleted',
+                ];
+            }
         } else {
-            return [
-                'message' => 'not deleted',
-            ];
+            return ['message' => 'unauthorised for this action'];
         }
     }
 }
